chore: move basicManifest fixture into the test helpers module
